#18 Move signup view to view folder, move image to assets/image folder
