+++ client/src/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
 import { Provider } from 'react-redux';
+import configureStore from './Store/configureStore';
+import App from './views/App';
 import * as serviceWorker from './serviceWorker';
 import {Route, Link, BrowserRouter as Router} from "react-router-dom";
-import Signup from './UserManagement/signup'
-import configureStore from './Store/configureStore';
+import Signup from './views/Signup'
 
 const store = configureStore();
